refactor: Playlist 모듈 구조 변경
