+++ playlist/data/src/main/java/kr/mastre/playlist/data/FirebasePlayItemDto.kt
@@ -1,7 +1,7 @@
 package kr.mastre.playlist.data
 
 import kotlinx.serialization.Serializable
-import kr.mastre.jaytv.playlist.domain.Playable
+import kr.mastre.playlist.Playable
 
 @Serializable
 internal data class FirebasePlayItemDto(val uri : String) : Playable {
