+++ playlist/data/src/main/java/kr/mastre/playlist/data/PlaylistRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package kr.mastre.playlist.data
 
 import io.reactivex.rxjava3.core.Single
-import kr.mastre.jaytv.playlist.domain.Playable
-import kr.mastre.jaytv.playlist.domain.PlaylistRepository
+import kr.mastre.playlist.Playable
+import kr.mastre.playlist.PlaylistRepository
 import javax.inject.Inject
 
 internal class PlaylistRepositoryImpl @Inject constructor(
