+++ playlist/src/main/java/kr/mastre/playlist/Playable.kt
@@ -1,4 +1,4 @@
-package kr.mastre.jaytv.playlist.domain
+package kr.mastre.playlist
 
 interface Playable {
     val rawUri: String
